# swagger.yaml
openapi: 3.0.0
info:
  title: API de Proyecto1
  description: API para gestionar usuarios (registro, login y recuperación de contraseña)
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /register:
    post:
      summary: "Registrar un nuevo usuario"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - name
                - email
                - password
      responses:
        '201':
          description: "Usuario registrado con éxito"
        '400':
          description: "Campos obligatorios no proporcionados"
        '500':
          description: "Error en el servidor"

  /login:
    post:
      summary: "Iniciar sesión"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: "Inicio de sesión exitoso"
        '400':
          description: "Correo y/o contraseña faltantes"
        '401':
          description: "Contraseña incorrecta"
        '404':
          description: "Correo no registrado"

  /recuperarpass:
    post:
      summary: "Recuperar contraseña"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '200':
          description: "Correo encontrado"
        '400':
          description: "Correo faltante"
        '404':
          description: "Correo no registrado"

  /actualizar-password:
    post:
      summary: "Actualizar contraseña"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                newPassword:
                  type: string
                repeatPassword:
                  type: string
              required:
                - email
                - newPassword
                - repeatPassword
      responses:
        '200':
          description: "Contraseña actualizada correctamente"
        '400':
          description: "Las contraseñas no coinciden"
        '404':
          description: "No se pudo actualizar la contraseña"
        '500':
          description: "Error en la base de datos"